add clause guards and log statements to controller file

+++ controllers/index.js
@@ -1,11 +1,17 @@
 import Profile from "../models/Profile.js";
 import ProfileRepo from "../repo/profileRepo.js";
 import ProfileService from "../services/ProfileService.js";
+import RequestType from '../utils/requestType.js'
 
 const profileRepo = new ProfileRepo(Profile);
 const profileService = new ProfileService(profileRepo);
 
 export const newProfile = async (req,res)=>{
+    if(!(req.body.request_type === RequestType.CREATE)){
+        res.send("Invalid Request");
+        return
+    }
+
     console.log("controller")
     try {
         const results = await profileService.createprofile(req);
@@ -17,6 +23,11 @@ export const newProfile = async (req,res)=>{
     }
 }
 export const fetchProfile = async (req,res)=>{
+    if(!(req.body.request_type === RequestType.GET)){
+        res.send("Invalid Request");
+        return
+    }
+
     console.log("FETCHING")
     try {
         const result = await profileService.getProfile(req);
@@ -26,6 +37,11 @@ export const fetchProfile = async (req,res)=>{
     }
 }
 export const updateProfile = async (req,res)=>{
+    if(!(req.body.request_type === RequestType.UPDATE)){
+        res.send("Invalid Request");
+        return
+    }
+    console.log("Updating...")
     try {
         const result = await profileService.updateProfile(req);
         res.json({data:result})
@@ -34,7 +50,12 @@ export const updateProfile = async (req,res)=>{
     }
 }
 export const deleteProfile = async(req,res)=>{
+    if(!(req.body.request_type === RequestType.DELETE)){
+        res.send("Invalid Request");
+        return
+    }
 
+    console.log("Delete")
     profileService.deleteProfile(req).then(result=> result && res.send("Successfuly deleted"))
     .catch(error=> console.log(error));
 }
